Updated index buffer size

+++ src/rendering/WorldGeometry.h
@@ -17,7 +17,7 @@ struct Vertex {
 };
 
 extern std::vector<Vertex> vertices;
-extern std::vector<uint16_t> indices;
+extern std::vector<uint32_t> indices;
 
 extern void addVerticesAndIndices(const std::vector<Vertex>& newVertices, const std::vector<uint16_t>& newIndices);
 extern void tesselateAndAddBlock(Block block);
